+++ SmartCity Core/src/main/java/be/uantwerpen/sc/models/Job.java
@@ -5,27 +5,27 @@ package be.uantwerpen.sc.models;
  */
 public class Job
 {
-    private int jobId;
+    private Long jobId;
     private String jobDescription;
 
     public Job(String jobDescription)
     {
-        this.jobId = 0;
+        this.jobId = 0L;
         this.jobDescription = jobDescription;
     }
 
-    public Job(int jobId, String jobDescription)
+    public Job(Long jobId, String jobDescription)
     {
         this.jobId = jobId;
         this.jobDescription = jobDescription;
     }
 
-    public int getJobId()
+    public Long getJobId()
     {
         return jobId;
     }
 
-    public void setJobId(int jobId)
+    public void setJobId(Long jobId)
     {
         this.jobId = jobId;
     }
