+++ SmartCity Core/src/main/java/be/uantwerpen/sc/services/PointControlService.java
@@ -21,7 +21,7 @@ public class PointControlService
         return pointRepository.findAll();
     }
 
-    public Point getPoint(int id)
+    public Point getPoint(Long id)
     {
         return pointRepository.findOne(id);
     }
